Leave only one test/index.ts that lets disabel/enable code coverage
This fix removes index.*.ts files and add index.ts file
that allows to run tests in with or without code coverage
reporting.
